login_button_with_contexts
Created a Contexts folder with a index file and Provider file so that I can store if/when the user is logged in and pass that to specific components down the tree if need be. For now I am wrapping the index component of core in the AuthorizationProvider element but may need to find a better way to do this in the future so I can pass multiple Contexts down.